Moved no-longer-used _format* methods to grammar_formatter.py.
